scripts/webvh-e2e/pyproject.toml
[tool.poetry]
name = "traction-webvh-e2e"
version = "0.1.0"
description = "End-to-end HTTP harness for did:webvh flows against the Traction tenant proxy"
authors = ["BC Government <digital.trust@gov.bc.ca>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.0"
python-dotenv = "^1.0.0"
click = ">=8.0,<9"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
